is_multimodal_model -> is_vision_model
