+++ autogen/oai/client.py
@@ -10,7 +10,7 @@ from flaml.automl.logger import logger_formatter
 from pydantic import BaseModel
 
 from autogen.cache.cache import Cache
-from autogen.img_utils import format_message_contents_with_images, is_multimodal_model
+from autogen.img_utils import format_message_contents_with_images, is_vision_model
 from autogen.io.base import IOStream
 from autogen.logger.logger_utils import get_current_ts
 from autogen.oai.openai_utils import OAI_PRICE1K, get_key, is_valid_api_key
@@ -162,7 +162,9 @@ class OpenAIClient:
         completions: Completions = self._oai_client.chat.completions if "messages" in params else self._oai_client.completions  # type: ignore [attr-defined]
         # If streaming is enabled and has messages, then iterate over the chunks of the response.
 
-        if is_multimodal_model(params.get("model", None)) and "messages" in params:
+        _is_vision = is_vision_model(params.get("model", None))
+        _is_vision |= params.get("vision-model", False)
+        if _is_vision and "messages" in params:
             params["messages"] = format_message_contents_with_images(params["messages"])
 
         if params.get("stream", False) and "messages" in params:
